requirements.txt
aniso8601==9.0.1
astroid==2.14.2
cached-property==1.5.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
dill==0.3.6
dnspython==2.3.0
email-validator==1.3.1
ez-setup==0.9
Flask==2.2.2
Flask-Login==0.6.2
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.1.1
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mongoengine==0.26.0
Pillow==9.4.0
platformdirs==3.0.0
pycparser==2.21
pygit2==1.11.1
pylint==2.16.2
pymongo==4.3.3
pypng==0.20220715.0
python-dotenv==0.21.1
pytz==2022.7.1
qrcode==7.4.2
six==1.16.0
smmap==5.0.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typed-ast==1.5.4
typing_extensions==4.5.0
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.0.1
uWSGI==2.0.21
